+++ openstl/datasets/dataloader.py
@@ -19,7 +19,7 @@ def load_data(dataname, batch_size, val_batch_size, num_workers, data_root, dist
     elif 'kth' in dataname:  # 'kth', 'kth20', 'kth40'
         from .dataloader_kth import load_data
         return load_data(batch_size, val_batch_size, data_root, num_workers, **cfg_dataloader)
-    elif dataname == 'mmnist':
+    elif 'mnist' in dataname:  # 'mmnist', 'mfmnist'
         from .dataloader_moving_mnist import load_data
         return load_data(batch_size, val_batch_size, data_root, num_workers, **cfg_dataloader)
     elif dataname == 'taxibj':