+++ openstl/datasets/dataloader_moving_mnist.py
@@ -10,39 +10,60 @@ from torch.utils.data import Dataset
 from openstl.datasets.utils import create_loader
 
 
-def load_mnist(root):
+def load_mnist(root, data_name='mnist'):
     # Load MNIST dataset for generating training data.
-    path = os.path.join(root, 'moving_mnist/train-images-idx3-ubyte.gz')
+    file_map = {
+        'mnist': 'moving_mnist/train-images-idx3-ubyte.gz',
+        'fmnist': 'moving_fmnist/train-images-idx3-ubyte.gz',
+    }
+    path = os.path.join(root, file_map[data_name])
     with gzip.open(path, 'rb') as f:
         mnist = np.frombuffer(f.read(), np.uint8, offset=16)
         mnist = mnist.reshape(-1, 28, 28)
     return mnist
 
-def load_fixed_set(root):
+
+def load_fixed_set(root, data_name='mnist'):
     # Load the fixed dataset
-    filename = 'moving_mnist/mnist_test_seq.npy'
-    path = os.path.join(root, filename)
+    file_map = {
+        'mnist': 'moving_mnist/mnist_test_seq.npy',
+        'fmnist': 'moving_fmnist/fmnist_test_seq.npy',
+    }
+    path = os.path.join(root, file_map[data_name])
     dataset = np.load(path)
     dataset = dataset[..., np.newaxis]
     return dataset
 
 
 class MovingMNIST(Dataset):
-    """Moving MNIST <http://arxiv.org/abs/1502.04681>`_ Dataset"""
+    """Moving MNIST Dataset <http://arxiv.org/abs/1502.04681>`_
 
-    def __init__(self, root, is_train=True, n_frames_input=10, n_frames_output=10,
-                 image_size=64, num_objects=[2], transform=None, use_augment=False):
+    Args:
+        data_root (str): Path to the dataset.
+        is_train (bool): Whether to use the train or test set.
+        data_name (str): Name of the MNIST modality.
+        n_frames_input, n_frames_output (int): The number of input and prediction
+            video frames.
+        image_size (int): Input resolution of the data.
+        num_objects (list): The number of moving objects in videos.
+        use_augment (bool): Whether to use augmentations (defaults to False).
+    """
+
+    def __init__(self, root, is_train=True, data_name='mnist',
+                 n_frames_input=10, n_frames_output=10, image_size=64,
+                 num_objects=[2], transform=None, use_augment=False):
         super(MovingMNIST, self).__init__()
 
         self.dataset = None
         self.is_train = is_train
+        self.data_name = data_name
         if self.is_train:
-            self.mnist = load_mnist(root)
+            self.mnist = load_mnist(root, data_name)
         else:
             if num_objects[0] != 2:
-                self.mnist = load_mnist(root)
+                self.mnist = load_mnist(root, data_name)
             else:
-                self.dataset = load_fixed_set(root)
+                self.dataset = load_fixed_set(root, data_name)
         self.length = int(1e4) if self.dataset is None else self.dataset.shape[1]
 
         self.num_objects = num_objects
@@ -65,7 +86,7 @@ class MovingMNIST(Dataset):
         x = random.random()
         y = random.random()
         theta = random.random() * 2 * np.pi
-        
+
         v_ys = [np.sin(theta)] * seq_length
         v_xs = [np.cos(theta)] * seq_length
 
@@ -179,16 +200,16 @@ class MovingMNIST(Dataset):
         return self.length
 
 
-def load_data(batch_size, val_batch_size, data_root, num_workers=4,
+def load_data(batch_size, val_batch_size, data_root, num_workers=4, data_name='mnist',
               pre_seq_length=10, aft_seq_length=10, in_shape=[10, 1, 64, 64],
               distributed=False, use_augment=False, use_prefetcher=False):
 
     image_size = in_shape[-1] if in_shape is not None else 64
-    train_set = MovingMNIST(root=data_root, is_train=True,
+    train_set = MovingMNIST(root=data_root, is_train=True, data_name=data_name,
                             n_frames_input=pre_seq_length,
                             n_frames_output=aft_seq_length, num_objects=[2],
                             image_size=image_size, use_augment=use_augment)
-    test_set = MovingMNIST(root=data_root, is_train=False,
+    test_set = MovingMNIST(root=data_root, is_train=False, data_name=data_name,
                            n_frames_input=pre_seq_length,
                            n_frames_output=aft_seq_length, num_objects=[2],
                            image_size=image_size, use_augment=False)
@@ -229,6 +250,7 @@ if __name__ == '__main__':
                   val_batch_size=4,
                   data_root='../../data/',
                   num_workers=4,
+                  data_name='mnist',
                   pre_seq_length=10, aft_seq_length=10,
                   distributed=True, use_prefetcher=False)