+++ openstl/datasets/dataset_constant.py
@@ -1,9 +1,17 @@
 dataset_parameters = {
+    'mfmnist': {
+        'in_shape': [10, 1, 64, 64],
+        'pre_seq_length': 10,
+        'aft_seq_length': 10,
+        'total_length': 20,
+        'data_name': 'fmnist',
+    },
     'mmnist': {
         'in_shape': [10, 1, 64, 64],
         'pre_seq_length': 10,
         'aft_seq_length': 10,
-        'total_length': 20
+        'total_length': 20,
+        'data_name': 'mnist',
     },
     'taxibj': {
         'in_shape': [4, 2, 32, 32],